isNumeric changed to isDecimal validation and added Yaxal as seeded user.

+++ server/migrations/20180725110753-create-plant-measurement.js
@@ -12,7 +12,7 @@ module.exports = {
         type: Sequelize.STRING
       },
       value: {
-        type: Sequelize.NUMERIC
+        type: Sequelize.DOUBLE
       },
       unit: {
         type: Sequelize.STRING
